Move tests into controller folders
Since they now run independently I needed to change the mongo adapter for oidc to reuse the db connection
