+++ src/controllers/oidc/adapter.ts
@@ -1,11 +1,14 @@
 import { HttpError } from '../../models/Error';
 import { snakeCase } from 'lodash';
-import { MongoClient } from 'mongodb';
-import { MONGODB_URI } from '../../util/secrets';
+import db from '../../util/database';
 
 const grantable = new Set(['access_token', 'authorization_code', 'refresh_token', 'device_code']);
 
 let DB: any;
+
+db.connection.once('open', async () => {
+    DB = db.connection.getClient().db();
+});
 
 class CollectionSet extends Set {
     add(name: string): any {
@@ -56,14 +59,6 @@ export default class MongoAdapter {
         this.name = snakeCase(name);
 
         collections.add(this.name);
-    }
-
-    static async connect() {
-        const client = new MongoClient(MONGODB_URI);
-        const connection: any = await client.connect();
-        const dbName = connection.s.options.dbName;
-
-        DB = connection.db(dbName);
     }
 
     coll() {
