--- test/e2e/_.test.ts
@@ -1,15 +1,0 @@
-import db from '../../src/util/database';
-import MongoAdapter from '../../src/controllers/oidc/adapter';
-
-beforeAll(async () => {
-    await MongoAdapter.connect();
-});
-
-afterAll(async () => {
-    await db.disconnect();
-});
-
-require('./account.ts');
-require('./grants.ts');
-require('./signin.ts');
-require('./signup.ts');
